Move PipeWire code to pipewire/ directory
